Add support for the JoyCaption model
This change errs on the side of following the example implementation
from the official Github in regards to model configuration/options.

Addresses #263

Adds support for JoyCaption using the newly processor. This model can be
used to caption many types of images in a variety of styles, such as
natural language, Booru tags, and more.

Moves usages of `torch.float16` to a class property `dtype` to allow for
this type to be overridden. The documentation states that JoyCaption
natively uses bfloat16 and I did not investigate if using float16 works.

JoyCaption project github: https://github.com/fpgaminer/joycaption.