Add export with the JPEG XL format
Advantages are smaller file sizes when compression is acceptable
(quality < 100) or lossless compression with quality = 100. Also the
alpha channel is supported and kept in the images.

Note 1: this only adds support for the export function, and is not general
JPEG XL support for taggui. There is the fork https://github.com/yggdrasil75/taggui
that does exactly this.

Note 2: You might need `pip install pillow-jxl-plugin` beforehand to
be able to export into JPEG XL.